Update e2e  tests to check chart data limits
